2100 Enable annotations within a fire boundary to be easily selected
Change the selection policy to only allow single annotation to be chosen
for edit tool.

So if user want to move or delete multiple annotations, please use
select tool. if user want to move or delete single annotations(for
example, choose a annotation in a Fire Boundary), please use edit tool.